+++ R/AllGenerics.R
@@ -925,7 +925,7 @@ setGeneric(".allowableXAxisChoices", function(x, se) standardGeneric(".allowable
 #' The default method will use all available (atomic) variables, but subclasses can specialize this to only allow, e.g., continuous or discrete variables.
 #' 
 #' @section Controlling hover choices:
-#' \code{.getTooltipUI(x, se, name)} should return an \code{HTML} tag definition that details information from selected columns in \code{rowData(se)} or \code{colData(se)} for a data point that is hovered.
+#' \code{.getTooltipUI(x, se, name)} should return an \code{HTML} tag definition representing information to display in the tooltip that is displayed in \code{DotPlot} panels when hovering over a data point.
 #' The data point is identified by \code{name}, its \code{rownames} or \code{colnames} value in \code{se}.
 #'
 #' @author Kevin Rue-Albrecht