+++ R/panelDefaults.R
@@ -68,6 +68,16 @@ panel.default.env$options <- list(
 #' \item \code{FontSize}, defaults to 1.
 #' \item \code{LegendPointSize}, defaults to 1.
 #' }
+#' 
+#' For \linkS4class{RowDotPlot}s:
+#' \itemize{
+#' \item \code{TooltipRowData}, defaults to \code{character(0)}.
+#' }
+#' 
+#' For \linkS4class{ColumnDotPlot}s:
+#' \itemize{
+#' \item \code{TooltipColumnData}, defaults to \code{character(0)}.
+#' }
 #'
 #' For \linkS4class{ComplexHeatmapPlot}s:
 #' \itemize{